without long long

+++ A_Add.c
@@ -9,7 +9,7 @@ int main()
 {
     int a, b;
     scanf("%d %d", &a, &b);
-    long long int total = add(a, b);
+    int total = add(a, b);
     printf("%d\n", total);
     return 0;
 }
